feat: add ModuleCard and ModuleModal components for module display and interaction
- Created ModuleCard.vue for displaying individual module cards with hover effects and selection functionality.
- Implemented ModuleModal.vue for showing detailed information about selected modules, including installation commands and features.
- Introduced index.ts to manage module data and exports for easier access.
- Added modules.css for styling related to module cards and animations.
- Refactored Index.vue to utilize new components, improving code organization and readability.
