+++ resources/js/components/ui/saucebase/modules.css
@@ -1,0 +1,103 @@
+.stripe {
+    --pattern-fg: color-mix(in oklch, var(--mod-color) 19%, transparent);
+    background-color: color-mix(in oklch, var(--mod-color) 7%, transparent);
+    background-image: repeating-linear-gradient(
+        -45deg,
+        var(--pattern-fg) 0px,
+        var(--pattern-fg) 1.5px,
+        transparent 1.5px,
+        transparent 5px
+    );
+    border: 1px solid color-mix(in oklch, var(--mod-color) 25%, transparent);
+
+    &::after {
+        content: '';
+        position: absolute;
+        inset: 0;
+        perspective-origin: bottom left;
+        width: 105%;
+        transform: translate(-4px, 6px) perspective(400px) rotateX(-25deg);
+        background-color: color-mix(
+            in oklch,
+            var(--color-foreground) 20%,
+            transparent
+        );
+        filter: blur(5px);
+        border-radius: 2rem;
+    }
+
+    .dark &::after {
+        background-color: color-mix(
+            in oklch,
+            var(--color-background) 50%,
+            transparent
+        );
+    }
+}
+
+/* Badge stripe uses foreground color regardless of module color */
+.badge-stripe {
+    --mod-color: var(--color-foreground);
+}
+
+@keyframes bounce {
+    0% {
+        transform: translate(0, 0);
+    }
+    25% {
+        transform: translate(-3px, 3px);
+    }
+    55% {
+        transform: translate(3px, -3px);
+    }
+    72% {
+        transform: translate(3px, -3px);
+    }
+    100% {
+        transform: translate(0, 0);
+    }
+}
+
+.bounce {
+    animation: bounce 0.7s ease-in-out backwards;
+}
+
+@keyframes stripe-appear {
+    from {
+        opacity: 0;
+        scale: 0.5;
+        transform: translateY(24px);
+    }
+}
+
+.stripe-appear {
+    transform-origin: left center;
+    animation: stripe-appear 0.6s ease-out backwards;
+}
+
+@keyframes card-drop-move {
+    0% {
+        transform: translateY(-100%) translateX(30%) scale(1.25);
+    }
+    100% {
+        transform: translateY(0) translateX(0) scale(1);
+    }
+}
+
+@keyframes card-drop-fade {
+    0% {
+        opacity: 0;
+    }
+    35% {
+        opacity: 1;
+    }
+    100% {
+        opacity: 1;
+    }
+}
+
+.card-drop {
+    animation:
+        card-drop-move 0.8s cubic-bezier(0.34, 1.56, 0.64, 1) both,
+        card-drop-fade 0.4s ease-out both;
+}
